# [ApiXmlAdjuster] Look for types in packages more efficiently.