When `JavaApiTypeResolverExtensions.FindNonGenericType()` is invoked, it searches through every type in every package to try to find a match.  This is ~5,000 types to search on each invocation for a default bindings project.

Since we have the full Java type name, like `java.lang.Object`, we can use that information to only look in packages that could potentially have the type:

* java
* java.lang

This helps speed up the `class-parse` case:

Scenario | Original | This PR
------------ | ------------- | -
class-parse | 13278ms | 7914ms