@jonpryor 2019-07-18
It's an interesting idea, but I'm not sure I "like" it.  (Which I suppose is beside the point, but still...)

Part of the problem is that the underlying API is using "Java" names.  The problem with "Java" names is that they're ambiguous around where the package ands and the class begins, which is evident with nested types.  Consider [`java.lang.Character.UnicodeBlock`](https://developer.android.com/reference/java/lang/Character.UnicodeBlock): `package_names` will contain `java` (good), `java.lang`, (good), and `java.lang.Character` (say what?).

This is *unavoidable* because `.` is used for both package name separation and inner type name separation; it's ambiguous.

Can `FindNonGenericType()` be updated to instead use *JNI* names?  These names are *unambiguous*, using `/` as the package separator and `$` as the nested type separator, thus `java/lang/Character$UnicodeBlock`.

Alternatively/Additionally, should `JavaApi.Packages` be a [`KeyedCollection<string, JavaPackage>`](https://docs.microsoft.com/en-us/dotnet/api/system.collections.objectmodel.keyedcollection-2?view=netframework-4.8) or `IDictionary<string, JavaPackage>`?  One can imagine a `JavaApi.Packages` implementation in which "normal" dictionary lookup can be made to hit both the "java" names *and* managed names, e.g. `api.Packages["android.view"]` will return the `Android.Views` namespace (for managed bindings), by using the `Android.Runtime.NamespaceMappingAttribute` values to suitably populate the mapping.

Inserting a dictionary should help *even more* than this PR; this PR still results in linear searches across `api.Packages`, it just helps restrict what the `pacakge.Types` lookup afterward.  If we could insert an O(1) dictionary lookup for package name *instead of* an O(n) package lookup, *surely* that would help *even more*?