@jpobst 2019-07-19
Can you elaborate more on your concern?  Is it that adding `java.lang.Character` to `package_names` is slightly less efficient?  (Note we will not search for types in it if there isn't actually a package for it.)  Or is there a correctness issue?  That is, will we not be resolving types correctly?

I had also spiked up using a `Dictionary` for `JavaApi.Packages`, which I agree should be faster.  However the speed difference was small enough that I couldn't measure it, so I went with this change as it is less invasive than changing the `JavaApi.Packages` type and fixing all the consumers.