@jonpryor 2019-07-19
> I had also spiked up using a Dictionary for JavaApi.Packages, which I agree should be faster. However the speed difference was small enough that I couldn't measure it,

Interesting.  Glad you looked at that.